Refactor main/driving code between VM and web apps
